lab3/rest_api/requirements.txt
djangorestframework==3.12.4
dj-database-url==0.5.0
Django==3.1.8
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.1.0